테스트 빌더 클래스 별도의 파일로 분리. (다른 테스트에서도 쓰고자 함)
